feat: Add html2canvas for element screenshotting
Integrates the `html2canvas` library to enable the functionality of capturing DOM elements as images. This will be used for generating screenshots of various application components, such as charts or reports, to facilitate export and sharing.

The `html2canvas` package has been added to dependencies and `index.html` has been updated to include it as a CDN dependency for browser compatibility.